Move IAM from Dialog to Activity
