+++ SwrveSDK/src/main/java/com/swrve/sdk/ISwrveBase.java
@@ -6,7 +6,6 @@ import android.content.Intent;
 
 import com.swrve.sdk.config.SwrveConfigBase;
 import com.swrve.sdk.messaging.ISwrveCustomButtonListener;
-import com.swrve.sdk.messaging.ISwrveDialogListener;
 import com.swrve.sdk.messaging.SwrveBaseCampaign;
 import com.swrve.sdk.messaging.ISwrveInstallButtonListener;
 import com.swrve.sdk.messaging.ISwrveMessageListener;
@@ -110,10 +109,6 @@ public interface ISwrveBase<T, C extends SwrveConfigBase> {
     ISwrveCustomButtonListener getCustomButtonListener();
 
     void setCustomButtonListener(ISwrveCustomButtonListener customButtonListener);
-
-    ISwrveDialogListener getDialogListener();
-
-    void setDialogListener(ISwrveDialogListener dialogListener);
 
     Context getContext();
