+++ SwrveSDK/src/main/java/com/swrve/sdk/SwrveImp.java
@@ -3,9 +3,8 @@ package com.swrve.sdk;
 import android.Manifest;
 import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.DialogInterface.OnDismissListener;
 import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
@@ -28,19 +27,14 @@ import com.swrve.sdk.localstorage.MemoryCachedLocalStorage;
 import com.swrve.sdk.localstorage.MemoryLocalStorage;
 import com.swrve.sdk.localstorage.SQLiteLocalStorage;
 import com.swrve.sdk.messaging.ISwrveCustomButtonListener;
-import com.swrve.sdk.messaging.ISwrveDialogListener;
 import com.swrve.sdk.messaging.ISwrveInstallButtonListener;
 import com.swrve.sdk.messaging.ISwrveMessageListener;
 import com.swrve.sdk.messaging.SwrveBaseCampaign;
-import com.swrve.sdk.messaging.SwrveCampaign;
+import com.swrve.sdk.messaging.SwrveInAppCampaign;
 import com.swrve.sdk.messaging.SwrveCampaignState;
 import com.swrve.sdk.messaging.SwrveConversationCampaign;
 import com.swrve.sdk.messaging.SwrveMessage;
 import com.swrve.sdk.messaging.SwrveOrientation;
-import com.swrve.sdk.messaging.view.SwrveDialog;
-import com.swrve.sdk.messaging.view.SwrveMessageView;
-import com.swrve.sdk.messaging.view.SwrveMessageViewBuildException;
-import com.swrve.sdk.messaging.view.SwrveMessageViewFactory;
 import com.swrve.sdk.qa.SwrveQAUser;
 import com.swrve.sdk.rest.IRESTClient;
 import com.swrve.sdk.rest.RESTClient;
@@ -113,15 +107,10 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
     protected static int DEFAULT_DELAY_FIRST_MESSAGE = 150;
     protected static long DEFAULT_MAX_SHOWS = 99999;
     protected static int DEFAULT_MIN_DELAY = 55;
-    protected static long MESSAGE_REAPPEAR_TIMEOUT = 1500;
-    protected static SwrveMessage messageDisplayed;
-    protected static long lastMessageDestroyed;
     private static String INSTALL_TIME_CATEGORY = "SwrveSDK.installTime";
-    protected final SimpleDateFormat timestampFormat = new SimpleDateFormat("HH:mm:ss ZZZZ", Locale.US);
     protected final SimpleDateFormat installTimeFormat = new SimpleDateFormat("yyyyMMdd", Locale.US);
     protected WeakReference<Context> context;
     protected WeakReference<Activity> activityContext;
-    protected WeakReference<SwrveDialog> currentDialog;
     protected String appVersion;
     protected int appId;
     protected String apiKey;
@@ -134,7 +123,6 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
     protected ISwrveConversationListener conversationListener;
     protected ISwrveInstallButtonListener installButtonListener;
     protected ISwrveCustomButtonListener customButtonListener;
-    protected ISwrveDialogListener dialogListener;
     protected ISwrveResourcesListener resourcesListener;
     protected ExecutorService autoShowExecutor;
     protected String userInstallTime;
@@ -178,8 +166,6 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
     protected String simOperatorCode;
     protected String androidId;
     protected int locationSegmentVersion;
-
-    protected int previousOrientation;
     protected SwrveQAUser qaUser;
 
     protected SwrveImp(Context context, int appId, String apiKey, C config) {
@@ -251,19 +237,6 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
             this.language = SwrveHelper.toLanguageTag(Locale.getDefault());
         } else {
             this.language = config.getLanguage();
-        }
-    }
-
-    protected void showPreviousMessage() {
-        if (config.isTalkEnabled()) {
-            // Re-launch message that was displayed before
-            if (messageDisplayed != null && messageListener != null) {
-                long currentTime = getNow().getTime();
-                if (currentTime < (lastMessageDestroyed + MESSAGE_REAPPEAR_TIMEOUT)) {
-                    messageListener.onMessage(messageDisplayed, false);
-                }
-                messageDisplayed = null;
-            }
         }
     }
 
@@ -676,7 +649,7 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
             if (messageListener != null && autoShowMessagesEnabled) {
                 SwrveMessage message = swrve.getMessageForEvent(SWRVE_AUTOSHOW_AT_SESSION_START_TRIGGER);
                 if (message != null && message.supportsOrientation(getDeviceOrientation())) {
-                    messageListener.onMessage(message, true);
+                    messageListener.onMessage(message);
                     autoShowMessagesEnabled = false;
                 }
             }
@@ -936,8 +909,8 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
         }));
     }
 
-    protected SwrveCampaign loadCampaignFromJSON(JSONObject campaignData, Set<String> assetsQueue) throws JSONException {
-        return new SwrveCampaign(this, campaignDisplayer, campaignData, assetsQueue);
+    protected SwrveInAppCampaign loadCampaignFromJSON(JSONObject campaignData, Set<String> assetsQueue) throws JSONException {
+        return new SwrveInAppCampaign(this, campaignDisplayer, campaignData, assetsQueue);
     }
 
     protected SwrveConversationCampaign loadConversationCampaignFromJSON(JSONObject campaignData, Set<String> assetsQueue) throws JSONException {
@@ -1032,18 +1005,9 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
         }
     }
 
-    protected void noMessagesWereShown(String event, String reason) {
-        SwrveLogger.i(LOG_TAG, "Not showing message for " + event + ": " + reason);
-        if (qaUser != null) {
-            qaUser.triggerFailure(event, reason);
-        }
-    }
-
-    protected void unbindAndShutdown(Activity callerActivity) {
+    protected void unbindAndShutdown() {
         // Reduce the references to the SDK
         int counter = bindCounter.decrementAndGet();
-
-        removeCurrentDialog(callerActivity);
 
         // Check if there are no more references to this object
         if (counter == 0) {
@@ -1051,50 +1015,6 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
             if (mustCleanInstance) {
                 ((SwrveBase<?, ?>) this).shutdown();
             }
-        }
-    }
-
-    protected void removeCurrentDialog(Activity callerActivity) {
-        if (currentDialog != null) {
-            final SwrveDialog dialog = currentDialog.get();
-            if (dialog != null && dialog.isShowing()) {
-                Activity dialogActivity = dialog.getOwnerActivity();
-                if (callerActivity == null || callerActivity ==  dialogActivity) {
-                    messageDisplayed = dialog.getMessage();
-                    lastMessageDestroyed = (new Date()).getTime();
-                    Activity activity = dialogActivity;
-                    if (activity == null) {
-                        activity = getActivityContext();
-                    }
-
-                    if (activity != null) {
-                        // Call from activity UI thread
-                        activity.runOnUiThread(new Runnable() {
-                            @Override
-                            public void run() {
-                                dialog.dismiss();
-                            }
-                        });
-                    } else {
-                        // Call from this thread
-                        dialog.dismiss();
-                    }
-                    currentDialog = null;
-                }
-            } else {
-                currentDialog = null;
-            }
-        }
-    }
-
-    protected void saveCurrentOrientation(Context ctx) {
-        try {
-            if (ctx != null) {
-                final Display display = ((WindowManager) ctx.getSystemService(Context.WINDOW_SERVICE)).getDefaultDisplay();
-                previousOrientation = display.getRotation();
-            }
-        } catch (Exception exp) {
-            SwrveLogger.e(LOG_TAG, "Could not obtain device orientation", exp);
         }
     }
 
@@ -1370,7 +1290,7 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
         }
     }
 
-    protected class DisplayMessageRunnable implements Runnable {
+    /*protected class DisplayMessageRunnable implements Runnable {
         private SwrveBase<?, ?> sdk;
         private Activity activity;
         private SwrveMessage message;
@@ -1435,5 +1355,5 @@ abstract class SwrveImp<T, C extends SwrveConfigBase> implements ISwrveCampaignM
                 SwrveLogger.w(LOG_TAG, "Couldn't create a SwrveMessageView", e);
             }
         }
-    }
+    }*/
 }
