+++ SwrveSDK/src/main/java/com/swrve/sdk/messaging/ISwrveMessageListener.java
@@ -10,8 +10,6 @@ public interface ISwrveMessageListener {
      * This method is invoked when a message should be shown in your app.
      *
      * @param message   message to be shown.
-     * @param firstTime indicates if this message was already showing and the app
-     *                  rotated.
      */
-    void onMessage(SwrveMessage message, boolean firstTime);
+    void onMessage(SwrveMessage message);
 }
